# workdir/docs/swagger.yaml
basePath: /
definitions:
  auth.Role:
    enum:
    - ADMIN
    - FORM_CREATOR
    - BASIC
    - BASIC
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleFormCreator
    - RoleBasic
    - DefaultRole
  errors.Code:
    enum:
    - invalid_argument
    - unauthorized
    - forbidden
    - not_found
    - conflict
    - internal
    type: string
    x-enum-varnames:
    - CodeInvalidArgument
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeInternal
  errors.HTTPErrorResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errors.Code'
        example: not_found
      message:
        example: form not found
        type: string
    type: object
...
  handlers.CreateSubmissionRequest:
    properties:
      completion_time_seconds:
        description: CompletionTimeSeconds is how long the respondent took to fill
          the form in.
        example: 95
        minimum: 0
        type: integer
      payload:
        description: Payload holds the answers; any JSON value but null.
        type: object
    type: object
...
  handlers.MetadataResponse:
    properties:
      completion_time_seconds:
        example: 95
        type: integer
      ip_address:
        example: 203.0.113.7
        type: string
      referer:
        example: https://example.com/contact
        type: string
      user_agent:
        example: Mozilla/5.0
        type: string
    type: object
...
  handlers.SampleResponse:
    properties:
      message:
        example: Hello, alice
        type: string
      policy:
        description: Policy describes who may call the endpoint.
        example: signed in
        type: string
      user:
        allOf:
        - $ref: '#/definitions/handlers.UserResponse'
        description: User is the signed-in caller; null for anonymous calls.
    type: object
...
  handlers.SubmissionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.SubmissionResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
    type: object
  handlers.SubmissionResponse:
    properties:
      form_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      metadata:
        $ref: '#/definitions/handlers.MetadataResponse'
      payload:
        type: object
      submitted_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        description: UserID is the signed-in submitter; null for anonymous submissions.
        example: 3
        type: integer
    type: object
...
  handlers.UserResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      role:
        allOf:
        - $ref: '#/definitions/auth.Role'
        enum:
        - ADMIN
        - FORM_CREATOR
        - BASIC
        example: BASIC
      username:
        example: alice
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Username and password of a user: "Basic " + btoa(username + ":" +
    password).'
  title: App API
  version: "1.0"
paths:
...
  /forms/{id}/submissions:
    get:
      description: Lists the submissions of one of the authenticated tenant's forms,
        newest first, with their metadata.
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      - default: 0
        description: Number of submissions to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: 20
        description: Maximum number of submissions to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Page of submissions
          schema:
            $ref: '#/definitions/handlers.SubmissionListResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List form submissions
      tags:
      - submissions
...
  /public/forms/{slug}/submissions:
    post:
      consumes:
      - application/json
      description: Stores the answers to a form that can currently be filled in, along
        with the client's IP address, user agent and referer. Forms that are not public_available
        require a signed-in user (bearer token or Basic credentials). The submitter's
        user ID is recorded only by forms that do not accept_anonymous.
      parameters:
      - description: Form slug
        in: path
        name: slug
        required: true
        type: string
      - description: Answers
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSubmissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Submission stored
          schema:
            $ref: '#/definitions/handlers.SubmissionResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Sign-in required, or invalid credentials
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: Form not found or not available
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Submit a form
      tags:
      - submissions
      - public
  /samples/admin:
    get:
      description: Requires the ADMIN role.
      produces:
      - application/json
      responses:
        "200":
          description: Greeting
          schema:
            $ref: '#/definitions/handlers.SampleResponse'
        "401":
          description: Not signed in
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "403":
          description: Role not allowed
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Admin sample
      tags:
      - samples
  /samples/form-creator:
    get:
      description: Requires the FORM_CREATOR role (ADMIN passes every role check).
      produces:
      - application/json
      responses:
        "200":
          description: Greeting
          schema:
            $ref: '#/definitions/handlers.SampleResponse'
        "401":
          description: Not signed in
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "403":
          description: Role not allowed
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Form creator sample
      tags:
      - samples
  /samples/public:
    get:
      description: Can be called by anyone. Credentials are optional, but rejected
        when invalid.
      produces:
      - application/json
      responses:
        "200":
          description: Greeting
          schema:
            $ref: '#/definitions/handlers.SampleResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      summary: Public sample
      tags:
      - samples
  /samples/signed-in:
    get:
      description: Requires a signed-in user of any role.
      produces:
      - application/json
      responses:
        "200":
          description: Greeting
          schema:
            $ref: '#/definitions/handlers.SampleResponse'
        "401":
          description: Not signed in
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Signed-in sample
      tags:
      - samples
...
securityDefinitions:
  ApiKeyAuth:
    description: API key of the tenant, returned when the tenant is created.
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: JWT returned by POST /auth/signin, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"